Have itemsets populate their own choice lists
