+++ src/main/java/org/javarosa/core/model/ItemsetBinding.java
@@ -7,11 +7,18 @@ import static org.javarosa.xpath.expr.XPathFuncExpr.toNumeric;
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
-
+import java.util.Map;
 import org.javarosa.core.model.condition.EvaluationContext;
 import org.javarosa.core.model.condition.IConditionExpr;
+import org.javarosa.core.model.data.IAnswerData;
+import org.javarosa.core.model.data.MultipleItemsData;
+import org.javarosa.core.model.data.StringData;
+import org.javarosa.core.model.data.helper.Selection;
 import org.javarosa.core.model.instance.DataInstance;
+import org.javarosa.core.model.instance.TreeElement;
 import org.javarosa.core.model.instance.TreeReference;
 import org.javarosa.core.model.util.restorable.RestoreUtils;
 import org.javarosa.core.services.locale.Localizable;
@@ -22,8 +29,11 @@ import org.javarosa.core.util.externalizable.ExtWrapNullable;
 import org.javarosa.core.util.externalizable.ExtWrapTagged;
 import org.javarosa.core.util.externalizable.Externalizable;
 import org.javarosa.core.util.externalizable.PrototypeFactory;
+import org.javarosa.debug.EvaluationResult;
+import org.javarosa.debug.Event;
 import org.javarosa.model.xform.XPathReference;
 import org.javarosa.xpath.XPathConditional;
+import org.javarosa.xpath.XPathException;
 import org.javarosa.xpath.expr.XPathNumericLiteral;
 import org.javarosa.xpath.expr.XPathPathExpr;
 import org.slf4j.Logger;
@@ -64,6 +74,121 @@ public class ItemsetBinding implements Externalizable, Localizable {
 
     public List<SelectChoice> getChoices () {
         return choices;
+    }
+
+
+    /**
+     * Creates a set of <code>SelectChoice</code> objects at the current question reference based on the data
+     * in the model.
+     * <p/>
+     * Always clears existing choices and repopulates when called.
+     */
+    public void populateDynamicChoices(FormDef formDef, TreeReference curQRef) {
+        formDef.getEventNotifier().publishEvent(new Event("Dynamic choices", new EvaluationResult(curQRef, null)));
+
+        List<SelectChoice> choices = new ArrayList<>();
+
+        List<TreeReference> matches = nodesetExpr.evalNodeset(formDef.getMainInstance(),
+            new EvaluationContext(formDef.getEvaluationContext(), contextRef.contextualize(curQRef)));
+
+        DataInstance formInstance;
+        if (nodesetRef.getInstanceName() != null) { // a secondary instance is specified
+            formInstance = formDef.getNonMainInstance(nodesetRef.getInstanceName());
+            if (formInstance == null) {
+                throw new XPathException("Instance " + nodesetRef.getInstanceName() + " not found");
+            }
+        } else {
+            formInstance = formDef.getMainInstance();
+        }
+
+        if (matches == null) {
+            throw new XPathException("Could not find references depended on by" + nodesetRef.getInstanceName());
+        }
+
+        // Get the answer to the current question to remove selection(s) that are no longer in the choice list.
+        Map<String, Boolean> currentAnswersInNewChoices = null;
+        IAnswerData rawValue = formDef.getMainInstance().resolveReference(curQRef).getValue();
+        if (rawValue != null) {
+            currentAnswersInNewChoices = new HashMap<>();
+
+            if (rawValue instanceof MultipleItemsData) {
+                for (Selection selection : (List<Selection>) rawValue.getValue()) {
+                    currentAnswersInNewChoices.put(selection.choice != null ? selection.choice.getValue() : selection.xmlValue, false);
+                }
+            } else {
+                currentAnswersInNewChoices.put(rawValue.getDisplayText(), false);
+            }
+        }
+
+        for (int i = 0; i < matches.size(); i++) {
+            TreeReference item = matches.get(i);
+
+            String label = labelExpr.evalReadable(formInstance, new EvaluationContext(formDef.getEvaluationContext(),
+                item));
+            String value = null;
+            TreeElement copyNode = null;
+            if (copyMode) {
+                copyNode = formDef.getMainInstance().resolveReference(copyRef.contextualize(item));
+            }
+            if (valueRef != null) {
+                value = valueExpr.evalReadable(formInstance, new EvaluationContext(formDef.getEvaluationContext(), item));
+            }
+
+            // Provide a default value if none is specified
+            value = value != null ? value : "dynamic:" + i;
+
+            if (currentAnswersInNewChoices != null && currentAnswersInNewChoices.keySet().contains(value)) {
+                currentAnswersInNewChoices.put(value, true);
+            }
+
+            SelectChoice choice = new SelectChoice(label, value, labelIsItext);
+            choice.setIndex(i);
+            if (copyMode)
+                choice.copyNode = copyNode;
+
+            choices.add(choice);
+        }
+
+        if (choices.size() == 0) {
+            logger.info("Dynamic select question has no choices! [{}]. If this occurs while " +
+                    "filling out a form (and not while saving an incomplete form), the filter " +
+                    "condition may have eliminated all the choices. Is that what you intended?"
+                , nodesetRef);
+
+        }
+
+        // Remove values that are no longer in choices.
+        if (currentAnswersInNewChoices != null && currentAnswersInNewChoices.containsValue(false)) {
+            IAnswerData filteredAnswer;
+            if (rawValue instanceof MultipleItemsData) {
+                filteredAnswer = getFilteredSelections((MultipleItemsData) rawValue, currentAnswersInNewChoices);
+            } else {
+                filteredAnswer = new StringData("");
+            }
+
+            formDef.getMainInstance().resolveReference(curQRef).setAnswer(filteredAnswer);
+        }
+
+        clearChoices();
+        setChoices(choices, formDef.getMainInstance(), formDef.getEvaluationContext(), formDef.getLocalizer());
+    }
+
+    /**
+     * @param selections          an answer to a multiple selection question
+     * @param shouldKeepSelection maps each value that could be in @{code selections} to a boolean representing whether
+     *                            or not it should be kept
+     * @return a copy of {@code selections} without the values that were mapped to false in {@code shouldKeepSelection}
+     */
+    private static MultipleItemsData getFilteredSelections(MultipleItemsData selections, Map<String, Boolean> shouldKeepSelection) {
+        List<Selection> newSelections = new ArrayList<>();
+        for (Selection oldSelection : (List<Selection>) selections.getValue()) {
+            String key = oldSelection.choice != null ? oldSelection.choice.getValue() : oldSelection.xmlValue;
+            if (shouldKeepSelection.get(key)) {
+                newSelections.add(oldSelection);
+            }
+        }
+
+        return new MultipleItemsData(newSelections);
     }
 
     private Long resolveRandomSeed(DataInstance model, EvaluationContext ec) {
