+++ src/main/java/org/javarosa/form/api/FormEntryPrompt.java
@@ -232,7 +232,7 @@ public class FormEntryPrompt extends FormEntryCaption {
         ItemsetBinding itemset = q.getDynamicChoices();
         if (itemset != null) {
             if (!dynamicChoicesPopulated) {
-                form.populateDynamicChoices(itemset, mTreeElement.getRef());
+                itemset.populateDynamicChoices(form, mTreeElement.getRef());
                 dynamicChoicesPopulated = true;
             }
             return itemset.getChoices();
